Split profile .h and .cc
